Data-Pipeline/requirements.txt
PyYAML==6.0.3
pandas==2.3.3
numpy==2.2.6
pytest>=8.0
python-dotenv==1.2.1
apache-airflow==3.1.7
